import coredns Deployment yaml
When we created the EKS cluster, it automagically creates a `coredns`
deployment within the cluster.

When you want to upgrade coredns, the documented way is to run

    kubectl set image --namespace kube-system deployment.apps/coredns \
      coredns=602401143452.dkr.ecr.us-west-2.amazonaws.com/eks/coredns:v1.3.1

:face_vomiting:

This commit imports the already-running coredns Deployment into the
cluster chart, so that we can upgrade it in a proper gitops manner
later.

It refers to a ServiceAccount `coredns` - this is also part of the
cluster and we should not import the yaml for that.  It has a
ClusterRole `system:coredns` and a ClusterRoleBinding
`system:coredns`, both of which are labelled
`kubernetes.io/bootstrapping: rbac-defaults` which means that they
belong to the cluster and you shouldn't manage them yourself.  See
https://kubernetes.io/docs/reference/access-authn-authz/rbac/#default-roles-and-role-bindings
for details.