import kube-proxy DaemonSet yaml
When we created the EKS cluster, it automagically creates a `kube-proxy`
daemonset within the cluster.

When you want to upgrade kube-proxy, the documented way is to run

    kubectl set image daemonset.apps/kube-proxy \
      -n kube-system \
      kube-proxy=602401143452.dkr.ecr.us-west-2.amazonaws.com/eks/kube-proxy:v1.13.10

:face_vomiting:

This commit imports the already-running kube-proxy DaemonSet into the
cluster chart, so that we can upgrade it in a proper gitops manner
later.

It refers to ConfigMaps `kube-proxy` and `kube-proxy-config`.  The
`kube-proxy` ConfigMap has cluster-specific config (ie the name of an
upstream EKS server to talk to) and so I don't think it's right to
pull this ConfigMap into our chart.  `kube-proxy-config` could go
either way - there's nothing unique to the specific cluster in there -
but I'm tempted to leave it out for consistency.  Opinions on this
welcome :)