Allow us to delete jobs
When a k8s Job hits backoffLimit, it will be marked as failed and any running
Pods will be terminated. This impedes our ability to debug it at a later date
as we won't be able to get logs etc.
With this commit we should be able to delete it and have it re-created by the
deployment pipeline.