vendor kubernetes-system rules from kubernetes-mixin
These rules come from
https://github.com/kubernetes-monitoring/kubernetes-mixin/ whose
README says

> This mixin is designed to be vendored into the repo with your
> infrastructure config.

We want to tweak some of these alerts.  I think as we find alerts we
want to tweak we should vendor them.  In each case we will need to
clean up the old PrometheusRule instances.  In this case, we will want
to:

    kubectl -n gsp-system delete prometheusrule gsp-prometheus-operator-kubernetes-system

I picked this file in particular because we want to shut up the
KubeVersionMismatch alert, which always alerts because the EKS plane
has a different point release to the nodes by design.