[KubeVersionMismatch alert] don't alert for mismatched point release
This changes KubeVersionMismatch so that it only alerts if we have
mismatched minor releases, not mismatched point releases.

Currently we have components on v1.13.8 (kubelet) and
v1.13.10 (apiserver).  The apiserver isn't in our control - it's
managed by EKS.  The kubelet is the version we get from the
EKS-provided AMI.

Therefore, it's by design that we have mismatched point releases at
the moment - we have the latest EKS-provided kubelet and the latest
EKS-provided control plane.

If we only care about minor releases, we can use the `major` and
`minor` labels instead of munging `gitVersion` with regular
expressions, which is another win.