fix alert templates
Helm is a bunch of go templates.

Prometheus alerts are a bunch of go templates.

Go templates do not nest well :(

I "fixed" this for now by wrapping each alert message that wants to
reference a variable in {{" "}} so that Helm just thinks it's a
literal string, so that the actual template can get through to
Prometheus.