remove broken scrape config
This is filling the prometheus logs with errors because
`/etc/istio-certs/root-cert.pem` doesn't exist.

This was copypasted from somewhere but we don't use it.  The intention
was that users can annotate their pods `prometheus.io/scrape: true` in
order to be scraped; this is generally not approved of these days (in
part because one pod might have multiple scrapable endpoints).  The
preferred solution is `ServiceMonitor` instead, which is what our
users in fact use.

As a result, this config block is safe to remove.