Add per-AZ node autoscaling groups
This adds new autoscaling groups for workers in each AZ.

But why tho?
------------

I want to enable the cluster autoscaler (see #244).  But the cluster
autoscaler can only scale with the tools that it has.  In particular,
if one AZ is under contention, it currently can't create a node in
that specific AZ.

What did you do?
----------------

I copied the `nodegroup.yaml` to a `nodegroup-v2.yaml`.  This is
similar to the original nodegroup except:

 - it uses launch templates instead of launch configurations

   - this is (in part) so that we can launch spot instances for
     workers in future

 - it doesn't create the NodeSecurityGroup (this is created in
   terraform instead)

 - it doesn't create the node target groups or node instance
   profiles (it steals them from the existing worker group instead)

 - it disallows outdated instance families in the NodeInstanceType
   field - no `m4` or `t2` instances for you :)

The security group change deserves more detail.  Previously, each node
group got its own security group, but the security groups were almost
identical to each other.  The only difference is that worker nodes
allow 31390 from public IPs, which ci and kiam nodes do not.

This does not scale to creating new worker nodegroups per AZ.  I think
a much simpler model would be to define a single `node` security group
which captures everything that a generic node needs, and if any extra
behaviour is needed (such as port 31390) this can be captured in an
extra security group (eg `worker` in this commit).

This means that, in future, all nodegroups should be in the same
`node` security group.  Particular nodegroups may wish to join other
security groups if they need particular access to extra things.

The new autoscaling groups will be scaled to 0 instances by default.
You can configure this with the `extra_workers_per_az_count` variable.
I'm going to raise a PR to set this to a reasonable value in sandbox
for testing, but I'd rather not create extra instances in verify or
portfolio just yet until we've demonstrated it working in sandbox.