introduces cluster "suspend" job
it would be nice to be able to put certain kinds of clusters (like
sandbox, test, or other ephemeral instances) into a kind of suspended
animation at night to save cash.

This introduces a "suspend" job that can be manually triggered that
cordons off and shutsdown ALL the worker intances.

Since the logic for shutting down workers is required my "destroy" job
too, I reused the same code in both jobs.