Document how to use tenant-provided TLS certificates
Now that we have per-namespace ingressgateways, it's possible for
tenants to provide their own TLS certificates.  This updates the
documentation to show how.

This also shows how to use a custom domain, provided the tenant
provides the certificate.

I tried as much as possible to show the details of exactly how to do
each step.  In particular, I wanted to show how to create the `Secret`
and to seal it, and I wanted to make it easy to do the right thing and
hard to accidentally commit an unsealed seacret.  Hence my slightly
funky bash pipeline to do it all in one line.

PR #599 is a separate approach which would allow tenants to use
cluster-provisioned certificates with custom domains.