Allow specifying workers to run on spot instances
This introduces a new cluster variable,
`worker_on_demand_percentage_above_min`, which defines the percentage
of instances that should be on-demand instances.  It defaults to 100.
But you can set it to (for example) 50, which means that the first
`minimum_workers_per_az_count` instances in each AZ nodegroup will be
on-demand, then half the workers beyond that will be on-demand and
half will be spot.

As a result, I expect this change to do nothing on its own, but we
could experiment with setting `worker-on-demand-percentage-above-min`
to different values in tech-ops-cluster-config.

Spot instances require supplying a list of allowed instance types to
use.  I picked instances based on the regex `[mr]5a?d?.xlarge`.  My
criteria were:

 - we don't need instance storage, but it won't hurt if we get some
 - we want instances to be at least `xlarge` so that we get 60 IPs per
   instance
 - we want the memory profile of at least `m5`, so the `c5` instances
   are ruled out
 - we want guaranteed compute capacity, so `t3` are ruled out
 - we don't care if we run Intel or AMD (I think?)

Much of the above is arguable - these aren't hard decisions but rather
just a way of picking something.

I think any of these instance types fit those criteria.  In terms of
on-demand pricing, `m5a` is cheapest and `r5d` is most expensive,
reflecting the fact that a) AMD is cheaper than Intel, and b) `m5` has
less memory than `r5`, and c) instance storage carries a premium.