Create gsp-critical PriorityClass
And get cluster-autoscaler to use it.

I've tested this by snowflaking in sandbox.

I think there may be a race condition here when spinning up a new
cluster: I don't know how to ensure the PriorityClass gets created
before the things that use it.

This is a proof of concept; we should have a conversation about which
things should run with this priority class (gatekeeper, the istio
sidecar injector, ...?).