run service operator tests in parallel
the controller tests are slow ... let's run them in parallel.

to prevent building the world N times, we build the base image, spawn N
tasks to run the tests and then copy the manager binary from the base image into
a scratch image for the final push.

we move to the registry-image / oci-build-task method to allow building
image and consuming in task without needing to push it up somewhere