add serial_group for cd-smoke-test
This is a workaround for a rare issue, that has started to show up more
since increasing the cd-smoke-test frequency.

we don't want the deploy job to run at the same time as the build job as
this increases the chance that we hit the "not signed by notary" bug if
a new build finishes before the previous deploy job has really begun.