use sed to avoid pipefail in harbor-resource
`head` will stop reading from it's stdin once it has happily done
it's job, leaving the stdin stream with unread data.

when used in a chain of processes piped together, leaving unread bytes
on the stream will cause a pipefail error to be raised in the recving
process.

This is all by design. Although the behaviour is somewhat surprizing.

The harbor resource sets `-o pipefail` to abort on any process that
returns a pipefail status (141). This helps us catch errors that occur
in chains of processes, without them being swallowed by the sending
process.

To parse the harbor project-name out of the full image uri, we were
spliting the uri into url segments, then using:
`echo $URI_SEGMENTS | head -n2 | tail -n1`
to grab the 2nd line of the file (the name).

I believe this is the culprit for the mysterious failures of the
resource we have seen.

Since the $URI_SEGMENTS string was so short, it is likely that this
almost always fit into the write buffer of the sending process (echo)
and so is instantly consumed in total by `head`... however sometimes
(totally up to the OS, hardware pressure, phase of the moon) the entire
string might not be sent/or read in one go... this would give `head` and
oportunity to stop reading more bytes, and ultimatly cause a pipefail
141 status.

To workaround this issue, I have switched to using `sed` to grab the 2nd
line of the URI_SEGMENTS and avoid the potential for pipefail.