disable concourse worker persistant volume
there is very little reason to keep the worker state in a persistant
volume when we explicitly tie one worker to one node.

we can use the "emptyDir" volume driver to provision a volume on the
node's local storage (which for m5d class instances is an SSD) which
will still give persistance until the node is terminated.

The main changes here should be that:

* volume/bagageclaim caches will no longer persist past node
  terminations. Personally I think this is fine. Builds logs and state
  are stored in postgres not on the workers so this will continue to be
  stateful.
* the "ci" instance type will dictate the available disk space (ie
  m5d.large is 75GB, m5d.xlarge is 150GB)

Additionally since we use "d" class instances in the production
clusters, I hope to see a speed up in disk heavy operations (like
starting tasks, container builds).