Fix default Gateway to exist if ingress is enabled
When I put this Gateway and Certificate in, I put them at the end of
the file without thinking too much.  Unfortunately I put them inside
the `if .repository` block, which mean that the canary namespace
didn't get a default Gateway.

This puts it in the correct place - inside the `if (default dict
.ingress).enabled` block, which means that we get a default Gateway if
and only if we have an ingressgateway to attach it to.