use clair to power a CVE smoke test tasl in conc
we'd like to know when we should be actioning upgrades of certain
components so we have confidence we are not leaving any nasty CVEs
unpatched, one quick idea for this would be to use clair which we
already have running in the cluster to report on any high CVEs from
running containers.

this uses klar (a cli for clair) installed into the task toolbox and
short script to fetch all running container images from the pod specs.