harbor resource: disbale content trust by default
we have yet to enforce signing of images using docker content trust
(notary) in the cluster and have decided we will be enforcing provence
of running images based on the chart's provenance and enforcing of use
of image digests (via gatekeeper).

we also believe that the content-trust/notary stage of builds is a
source of intermitant failures as it appear to be quite sensitive to
timeouts / load / etc.

the combination of the two points above means we no longer want to have
content trust on my default in the harbor resource, and so we set it to
false instead.

Anyone not explicitly setting `enable_content_trust: true` will no
longer be signing their images and their harbor image will stop
rejecting pulls when images are not signed