Fix Terraform plan for a new cluster
We are currently running into the following Terraform bug when trying to apply
a new cluster using Terraform 0.11:
https://github.com/hashicorp/terraform/issues/12570#issuecomment-309995500
This is because we are trying to run `length()` on a thing that references
another module which has not yet been applied.
We know what the length of it is already, so we can just output the
pre-calculated value.