Default to using the internal registry API
This is a very specific to GSP change that means we can remove the API
endpoints from the internet-facing ingress for Harbor. This means we can
then remove the IP safelist for those endpoints because they will only
be available internally anyway.

This change (and the associated work to remove the API endpoint routes
from the Harbor `VirtualService`) assumes that nothing is using Harbor's
API other than this resource.