limit concourse worker disk to 60Gi
Limit the empty volume (instance storage) concourse workers are using
for there ephemeral storage to 60Gi so that they don't fill up the
node's disk.

We believe this being unlimited to be a source of previous disk-pressure
and ultimate responsible for the scheduler eviciting important pods
(like CNIs and kube-proxy) to releave that pressure.

The number 60Gi was chosen as we are giving the ci nodes a 75Gi disk and
want to leave plenty of headroom for the OS and anything else on the
node.

In a future kubernetes version node-local storage is an option for
persistant volumes, which may be a more practical solution than limiting
the emptyVolume size (which doesn't appear to play well with how the
scheduler decides how to manage evicitions due to disk-pressure).