Use the internal service to access Harbor API
We want to be able to remove the client IP safelist. We've decided to do
this by removing the need to have an ingress for the API entirely. This
change will mean that Concourse will use the internal service for the
Harbor API because the `concourse-harbor-resource` will use it by
default.