use externalTrafficPolicy: Cluster for ingress LBs
sets the managed namespace loadbalaners to use the "cluster" mode (the
default mode). In this mode, loadbalancers will include all node in
there target groups and will rely on the inter-cluster networking to
route traffic to the correct pod, instead of directly targeting the node
where pod is currently located. This looks less effcient, but is less
sensitive to node failure.

we previously used externalTrafficPolicy: Local as we believed we had a
need to preserve client ip addrs to perform whitelisting tasks.

we have since discovered that we can relax this need and have removed
our use of ip whitelisting of resources from the system.

we suspect having the ingress loadbalancer service in
`externalTrafficPolicy: Local` mode has contributed to problems we have
seen during node rollouts (slower to update target groups on loss of a
node) and control plane upgrades (load balancers in a poor state after
upgrade - potentially as this configuration is less common and less
tested by the masses).