allow scaling ingressgateways to 2
Allow running a minimum of 2 ingressgateway pods instead of 3.

We previously had to ensure that there was always 3x ingressgateways
spun up and distributed over the three AZs. This was due to a
combination of how externalTrafficPoliicy:Local and the lack of
cross-zone loadbalancing from the provisioned NLBs would interact.

We no longer use externalTrafficPoliicy:Local and our NLBs are
configured to do cross-zone load balancing, so neither of these
conditions apply and we are safe to allow scaling down to two instances
of the ingressgateway pods. (we still want at least two though for HA).