remove harbor ip safelist rules
Removes the istio policy rules that prevent access to the harbor api
from anything other than the specified safelist of ips (which was only
the cluster's own egress ips).

The need for this policy has since been replaced by targeting the harbor
api only via the cluster's internal private address and not exposing it
to the world via a public ingress at all.