ensure canary deploys from a trusted release
make the cd-smoke-test pipeline use our github-resource (so that it gets
tested along with the secrets/authentication that we expect to exists in managed
namespaces) and set the branch/tag that the resource watches to be the same
as the last release that was built (so that canary code that has not
passed the release process does not get deploed into clusters)

The default cluster.release value is set to "master" in case the chart is
rendered outside of a release.

We set the approval count to zero since different cluster dpeloyments
have different trusted developers, instead we trust that
((cluster.release)) is set correctly.

Closes #214