make everyone an Editor in grafana org=1
allow everyone to edit dashboards in grafana.

even though we do not persist grafana configuration (prefuring to setup
dashboards declaratively) it is still useful to be able to click around
and make changes in the UI to test out dashboard changes.

the viewers_can_edit permisisons only gets us so far, but is missing
some key editing features like "creating a dashboard"