ensure all loggroups recv traffic in check-logging
we recently changed the name of the log group where contianer logs get
shipped to in cloudwatch, as well as adding some additional log shipping
for things running on the worker node's themselves.

this updates the check-logging test in the pipeline to watch for logs
arriving in all of these groups.