build(svcop): use GOPROXY to fetch go modules
This sets the GOPROXY environment variable to the default version in
go 1.13.  This means we fetch modules from the golang.org go proxy,
rather than direct.  This would have avoided the issues we saw with
honnef.co the other day (so I have reverted that change).  I also
think that it may make the build faster on concourse; it's worth a go.

In any case, when we upgrade to go 1.13, this will be the default
setting so we might as well try it out now.