Add control-plane label to kube-system.
If a node is removed that carries the only running instance of a webhook
target without a notification to the apiserver the control-plane
effectively breaks because the webhook is called as part of controller-
manager calls into the apiserver. Excluding the kube control-plane from
such checks allows the cluster to continue to function.