CloudForm a default desired worker node count of maximum
We think that when the launch template changes CloudFormation resets the
desired count away from whatever the autoscaler left back to what we put in
CloudFormation. This leads to the cluster scaling down to the minimum size as
old nodes are killed and replaced only with the minimum amount.
The cluster should heal itself when (if) the autoscaler comes back up, but we
may prefer a situation where it creates more than it needs and then the
autoscaler is responsible for removing unnecessary nodes.