use attempts to retry check-logging tasks
we often see the check-logging jobs fail to retry correctly.

without looking too deeply at the job the likely culprits are (1) the
retry logic is not quite right and (2) 'jq | head' is a known pipefail
scenario.

so as an attempt at a quick fix for the flakey test, (1) use concourse
'attempts' to retry instead of a bash loop and (2) drop the error on
pipefail option and live on the stream-redirecting-edge