reduce time waiting between node launches to 2 minutes
We currently wait 5 minutes after terminating and launching an
instance in the worker node group ASG.

We now have lifecycle hooks to ensure we are confident that terminated
instances have successfully evicted their pods onto other nodes.

I think we could *possibly* get rid of PauseTime entirely, but this is
a more conservative proposal for now.