Set log forwarding runtime to python3.7
I built the ZIP in #847 from a Debian Buster VM instead of something else like
Ubuntu Bionic, and the default version of Python 3 was 3.7 rather than 3.6.
This should be fine. However it does seem to mean we need to bump the version
of python used to run the lambda.

Alternatively I could make the ZIP again using a distro with Python 3.6 by
default, or attempt to get Python 3.6 on Debian Buster and make the ZIP
build process use that.