Implement roles as per ADR043
Following an audit of resources in a GSP deployment and with "dev
namespaces" in mind ADR043 describes the permissions that were agreed
on. This implements those roles and the necessary plumbing to wire them
in correctly.

Also included is some tidying up of legacy config relating to old roles
and users from IAM. The deployment pipeline also includes code to remove
the old roles and role bindings.