remove cd-smoke-test serial group
the serial-group, preventing the "build" and "deploy" jobs of the
cd-smoke-test pipeline running at the same time, was added to work
around the notary-only-signs-latest-tag bug. This bug is no longer a
factor as we are not signing the images with notary, so it's fine to
deploy old versions of images and we can remove the serial group.