stop dataplane, host, legacy log ship to splunk
The (legacy) "pods" log group is no-longer used, and this is just dead
code while we waited for the transition from "pods" -> "application" log
group... so remove it!

The "database" and "host" log groups are mostly unstructured logs that
are currently not used in any splunk protective monitoring and so is
just contributing to increased log volume costs and noise in splunk
queries .... so remove them too!

"application" (pod logs) and "eks" (audit/schduler/api logs) required in
many protective monitoring use-cases continue to be shipped unaffected.