allow config.istio.io Handler kubeyaml
This adds Handlers to the list of things that can be deleted by
Auditors and created/patched/updated by Operators.

We already do this with the closely-related Instances, Rules, and
Metrics; Handlers should be subject to the same access control.

This is used by the proxy node to add some custom Istio metrics.