Remove ingressgateway HorizontalPodAutoscaler
We are causing alert noise in tenant namespace with KubeHpaMaxedOut
alerts.  This is because ingressgateways have a HPA with min = max =
3.

This commit sets `autoscaleEnabled: false` which will remove the HPA
kubeyaml.  We might need to go and manually remove the HPA objects
afterwards.

We already set `replicaCount: 3` so the Deployment should have the
correct number of replicas without `autoscaleEnabled: true`.