Remove deprecated log group
We have a log group with the same name as the cluster domain.  We no
longer write any logs to it, despite setting the fluentd-cloudwatch
`logGroupName` value to it.

This is because we override the log group name on an output-by-output
basis in the fluentd configuration file, so nothing goes to this
top-level log group.

The default value for `logGroupName` is `kubernetes` which is a log
group which doesn't exist.  Thought: maybe the default should be one
of the log groups which does exist?  In any case, it shouldn't be the
old and deprecated one.