switch canary to use ImageRepository for image
we will be dropping support for the in-cluster harbor registry soon, so
we want to switch our cd-smoke-test to using the new ImageRepository
(ECR) based image storage.

An ImageRepository (ECR repository) and Prinicpal (IAM Role) are created
for the canary, which will populate a `Secret` called "canary" in the
"xxxxx-main" namespace. Once provisioned that secret will contain the
creds and URI required to push to the repository.