update splunk forwarder lambda to fix host field
splunk log events have a "host" field that defaults to the log-group
name, some events attempt to set this field to a namespace/pod string so
that events can be filtered by namespace.

This version of the log shipping lambda[1] resolves an issue that is
preventing the "host" field being populated as expected.

[1] https://github.com/alphagov/cyber-cloudwatch-fluentd-to-hec/pull/21