default to NOT checking creds on pipeline apply
the concourse-operator currently performs set-pipeline with "check
creds" flag set ... this means that an Pipeline resource will fail to
reconcile if it uses secret variables that are not yet defined.

This behaviour is undesirable since nobody can see the reconcile error
unless they inspect the operator's logs or note the Pipeline status.

Instead, just let the operator set the pipeline, and the issues will
become visible in the concourse ui