canary: ecr based cd-smoke-test
We have switched to using service-operator provisioned ImageRepositories
(ECR) as the supported image storage.

Unfortunatly the docker-image resource makes some weird assumptions when
the registry uri "looks like" an ECR URI, triggering a code path that
want an AWS access key/id instead of username/password, that is
incompatible with us already knowing the username/password.

The newer builtin registry-image[1] resource does not suffer from the
same issue, but also requires us to seperate out the "build" step to
as registry-image is only responsible for pushing/pulling from
registries.

This change updates the cd-smoke-test pipeline to use the
oci-build-task[2] to build the canary, and the registry-image resource
to store the image.

[1] https://github.com/concourse/registry-image-resource
[2] https://github.com/vito/oci-build-task