enforce digest use for all images in account's ecr
we currently enforce that pods can only reference images from the local
harbor registry if they use a digest. this improves our confidence that
what is running in the cluster is what we think it is, avoiding tag
mutability issues.

this adds the same restriction for images referenced from the account's
ecr registry.