Isolate tenant namespace istio resources
Istio violates the cluster-scoped vs namespace-scoped separation with
resources such as ServiceEntry that could allow a resource created in
one namespace to affect the state of the whole cluster. This uses
gatekeeper to enforce namespace isolation for these offending istio
resource types.

Some namespaces (kube-system, gsp-system, istio-system) are excluded as
some entries need to be global (such as the egress safelist configured
in *-cluster-config).