Fix external dns
A bug was introduced whereby newly-deployed external-dns instances will
fail to start up (discovered on an on-demand cluster). The problem seems
to be related to the way the RoleBinding is set up. Currently the
RoleBinding is in the wrong namespace (it's being created in gsp-system
instead of the tenant namespace). It makes sense that breaks it.

However, adding an equivalent RoleBinding to the correct namespace with
the same values otherwise doesn't fix external-dns. It has to be a
ClusterRoleBinding. Why this is is unclear to me at this stage.

Fixing it in this way doesn't break dev-namespace isolation and, as far
as I can tell, doesn't introduce any security problems. A user of a
namespace (including in-cluster concourse) cannot create cluster-wide
resources. These ClusterRoleBindings are being created by the
multi-tenant concourse.