svcop: add a default ecr lifecycle policy
there is a limit to the number of container images each ECR repository
can store.

This adds an image lifecycle policy[1] to service operator provisioned ECR
repositories that will only keep a history of the last 100 most recent
images. 100 was chosen simply for being 10x less than the 1000 image
maximum limit, so if there is a better default for the future use that.

[1] https://docs.aws.amazon.com/AmazonECR/latest/userguide/LifecyclePolicies.html