Allow all nodes to read CloudWatch metrics
We are planning on removing Kiam. It is being replaced by IAM roles for
service accounts on Amazon EKS clusters[1].

Grafana is the only thing in GSP that still uses Kiam. We are waiting
for an upstream fix[2] in order to use IAM roles for service accounts.
It doesn't appear that this will be resolved any time soon. In order to
allow Grafana to read CloudWatch metrics once Kiam has been removed
we've decided to allow _all_ nodes (and therefore _all_ containers) to
be able to read CloudWatch metrics.

[1] https://docs.aws.amazon.com/eks/latest/userguide/iam-roles-for-service-accounts.html
[2] https://github.com/grafana/grafana/issues/20473#issuecomment-559181587