Block all non-`gsp-system` namespaces from the AWS metadata service
This change grants access to the metadata service to pods in
`gsp-system` iff they have the `talksToAWSMetadataService: "true"`
label.

Tenants do not need access to this--they should be using IAM roles for
Service Accounts.

The `cluster-autoscaler` requires access to pull it's access token.
This could probably be changed to make it worth with IAM roles for
Service Accounts.

Grafana needs this to pull access tokens because it doesn't currently
worth with IAM roles for Service Accounts.