add "generation" label to ci worker nodes
this add a "generation" label to ci (concourse) worker nodes that gets
populated with a timestamp from the deployment pipeline.

a timer in the deployment pipeline will trigger daily, causing the
generation label to be updating in the nodegroup configuration and cause
cloudformation to roll the nodes via the same mechanism that occurs with
any other node modification.

if useful, this could easily be added to other node groups, but right
now we just want the concourse workers to get kicked at least once a
day to avoid them getting a case of bloaty-cache