add "generation" param to force node group update
This add a "NodeGroupGenerationTimestamp" cloudformation param to node
groups with a value populated with a timestamp pulled in from the
deployment pipeline.

The intention is to enable periodic rolling of nodes.

A timer in the deployment pipeline will trigger daily, causing the
generation timestamp to be updated in the nodegroup configuration and
cause cloudformation to roll the nodes via the same mechanism that
occurs with any other node modification.

The NodeGroupGenerationTimestamp does not actually _do_ anything, it
is simply interpolated into the userdata script in a harmless `echo`.

If useful, this could easily be added to other node groups, but right
now we just want the concourse workers to get kicked at least once a
day to avoid them getting a case of bloaty-cache

Note for travelers from the future: Please remove this hack and instead
take advantage of ASG Maximum Instance Lifetime[1]. This festure was
unavailable in cloudformation at the time of writing[2]

[1] https://aws.amazon.com/about-aws/whats-new/2019/11/amazon-ec2-auto-scaling-supports-max-instance-lifetime/<Paste>
[2] https://github.com/aws-cloudformation/aws-cloudformation-coverage-roadmap/issues/284